Added conditional renders for tabs and pages
Specific tabs and pages will only render depending on the type of the
authenticated user. Attempts to manually type the URL and enter an
unauthorized page will kick the user to the dashboard, which is a
global page across all user types.

Future revamp of this feature may employ JWT manipulation.