Removed levels from engaged subjects
Felt like it wasn't really necessary to include level of subject in
engaged subjects as there isn't much you can do with it. Also causes
some symmetry issues when it comes to the contractual interface, as the
student would not be able to select the level of the subject; tutors do
not have the option to select subject levels.