Change announcement display and delete logic
Display is now more fluid and updates instantaneously upon component
mount; that is, the window does not need to refresh to update
announcements. Same applies for announcement deletions; announcements
now have a unique id that can be used to more readily delete
announcements. This also resolves the issue where it was impossible to
differentiate announcements with same title and body.